Refactor configuration classes
