+++ object-storage-aws/src/main/java/io/micronaut/objectstorage/aws/AwsS3Configuration.java
@@ -19,8 +19,8 @@ import io.micronaut.context.annotation.EachProperty;
 import io.micronaut.context.annotation.Parameter;
 import io.micronaut.core.annotation.Introspected;
 import io.micronaut.core.annotation.NonNull;
-import io.micronaut.objectstorage.AbstractObjectStorageConfiguration;
-import io.micronaut.objectstorage.ObjectStorageConfiguration;
+import io.micronaut.objectstorage.configuration.AbstractObjectStorageConfiguration;
+import io.micronaut.objectstorage.configuration.ObjectStorageConfiguration;
 
 import javax.validation.constraints.Pattern;
